+++ tests/test_cli.py
@@ -244,6 +244,7 @@ class CliTestCase(TestCase):
                         skip_created_before=None,
                         skip_created_after=None,
                         skip_photos=False,
+                        no_smart_albums=False,
                     ),
                     UserConfig(
                         directory="def",
@@ -284,6 +285,7 @@ class CliTestCase(TestCase):
                         skip_created_before=None,
                         skip_created_after=None,
                         skip_photos=False,
+                        no_smart_albums=False,
                     ),
                 ],
             ),
@@ -364,6 +366,7 @@ class CliTestCase(TestCase):
                         ),
                         skip_created_after=datetime.timedelta(days=2),
                         skip_photos=False,
+                        no_smart_albums=False,
                     ),
                 ],
             ),